refactorings, code structure
