+++ tests/test_zoned_datetime.py
@@ -2545,6 +2545,7 @@ class TestRound:
     @pytest.mark.parametrize(
         "unit, increment",
         [
+            # TODO zero, negative, and floats
             ("minute", 8),
             ("second", 14),
             ("millisecond", 15),
